Require Enter when run key fails due to nearby enemies
It was really easy to press the direction automatically after pressing
the run key, even if the run prompt didn't show up due to the presence
of visible enemies.

This commit requires that Enter (or the custom prompt dismissal key) be
pressed if the run key fails due to visible enemies.