Selecting ground/inventory ammo reloads wielded weapon
A wielded gun can be reloaded with ammo off the ground (with the use
key) or in the inventory (by selecting it).  This provides a way to do a
normal reload that ignores equipped ammo boxes without having to unequip
it first.

Inspired by a forum post that mistakenly believed that this feature
already existed in the game.