Separate CF_OVERWALL instead of trying to use CF_STICKWALL and CF_OPENABLE.
In the current cells, the old behavior was wrong for locked doors, and levels had to jump through hoops to get it right.

(Nothing uses CF_OVERWALL, but it seems like it could be useful.)

Inferno also had to do a lot of work-arounds for this, so I think it is good to be flexible.